refactor: fix P0 circular dep, SPARQL injection, and split server monolith
- P0-1: Break circular dependency between snapshot.ts and server.ts
  by importing persistGraph from persist.ts instead of server.ts
- P0-2: Fix SPARQL injection in handleContextSearch and
  handleContextImpact using escapeTurtleLiteral()
- P0-3: Extract 1938-line server.ts into 4 handler modules:
  handlers/rdf.ts (RDF lifecycle),
  handlers/system.ts (status, graph CRUD, rollback),
  handlers/schema.ts (schema, visualize, infer),
  handlers/context.ts (6 context handlers)
  Server.ts reduced to slim 752-line entry point with dispatcher
